upgrade TELR liftover strategy, expand final report
